# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JOIN API",
    "description": "API мини-приложения **JOIN** в MAX: афиша города, события по Пушкинской карте и поиск компании, чтобы пойти вместе.\n\n### Основной сценарий\n1. **Вход** — `POST /api/auth/login` (или регистрация) возвращает `token`.\n2. **Афиша** — `GET /api/events?city=Казань`, фильтр `pushkinCard=true`.\n3. **Сохранить событие** — `POST /api/events/{id}/like`. Если его уже сохранил подходящий человек, в ответе будет совпадение.\n4. **Позвать пойти вместе** — `POST /api/matches/{id}/request`, второй отвечает `POST /api/matches/{id}/accept`. Только после этого открывается чат.\n5. **Чат** — `GET` / `POST /api/chats/{chatId}/messages`.\n\n### Как попробовать\nПолучите токен через `POST /api/auth/login`, нажмите **Authorize** и вставьте его в `bearerAuth`. Внутри MAX вместо токена приложение передаёт заголовок `X-Max-Init-Data`.\n\n### Ответы\n* Без авторизации защищённые методы отвечают `403` с пустым телом.\n* Ошибки — JSON `ErrorResponse`: `{\"error\": \"текст для пользователя\", \"code\": \"…\"}`.\n* Списки с `page`, `size`, `sort` возвращают страницу: `content`, `totalElements`, `number`, `size`.\n* Даты — `YYYY-MM-DD`, время — `HH:mm:ss`, местное время города события.\n\nОбязательные проверки основного сценария — `DATA-API.yaml` в репозитории.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://2-27-20-14.sslip.io",
      "description": "Демо-стенд"
    },
    {
      "url": "/",
...
    }
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "maxInitData": []
    }
  ],
...
  "paths": {
...
    "/api/support/ticket/read": {
      "put": {
        "tags": [
          "Поддержка"
        ],
        "summary": "Отметить ответы поддержки прочитанными",
        "operationId": "markRead",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/support/ticket/messages": {
      "get": {
        "tags": [
          "Поддержка"
        ],
        "summary": "Переписка с поддержкой",
        "operationId": "getMessages",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSupportMessageResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Поддержка"
        ],
        "summary": "Написать в поддержку",
        "operationId": "sendMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendSupportMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportMessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "post": {
        "tags": [
          "Уведомления"
        ],
        "summary": "Отметить уведомление прочитанным",
        "operationId": "markRead_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/matches/{id}/request": {
      "post": {
        "tags": [
          "Совпадения и приглашения"
        ],
        "summary": "Позвать пойти вместе",
        "description": "Второй получит приглашение в приложении и сообщение бота с кнопками «Пойдём» / «Не в этот раз». Чата ещё нет.",
        "operationId": "request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Второй уже ответил на приглашение",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/group-chats/{groupChatId}/messages": {
      "get": {
        "tags": [
          "Компании на событие"
        ],
        "summary": "Сообщения чата компании",
        "operationId": "getMessages_1",
        "parameters": [
          {
            "name": "groupChatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageGroupChatMessageResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Компании на событие"
        ],
        "summary": "Написать в чат компании",
        "operationId": "sendMessage_1",
        "parameters": [
          {
            "name": "groupChatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendGroupMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupChatMessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/friend-groups/{groupId}/messages": {
      "get": {
        "tags": [
          "Группы друзей"
        ],
        "summary": "Сообщения группы друзей",
        "operationId": "getMessages_2",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageFriendGroupChatMessageResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Группы друзей"
        ],
        "summary": "Написать в группу друзей",
        "operationId": "send",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendGroupMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendGroupChatMessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/chats/{id}/messages": {
      "get": {
        "tags": [
          "Личные чаты"
        ],
        "summary": "Сообщения чата",
        "description": "Новые сообщения также приходят через WebSocket `/ws`.",
        "operationId": "getMessages_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageChatMessageResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Личные чаты"
        ],
        "summary": "Отправить сообщение",
        "description": "Собеседник получит его через WebSocket и уведомление бота.",
        "operationId": "sendMessage_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Один из собеседников в чёрном списке у другого",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tags": {
      "get": {
        "tags": [
          "Афиша"
        ],
        "summary": "Поиск тегов",
        "description": "По подстроке `query`; без неё — популярные теги.",
        "operationId": "searchTags",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagResponse"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/tags/popular": {
      "get": {
        "tags": [
          "Афиша"
        ],
        "summary": "Популярные теги",
        "description": "Десять тегов с наибольшим числом событий.",
        "operationId": "getPopularTags",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagResponse"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/support/ticket": {
      "get": {
        "tags": [
          "Поддержка"
        ],
        "summary": "Обращение в поддержку",
        "description": "Возвращает обращение пользователя; при первом вызове создаёт его.",
        "operationId": "getOrCreateTicket",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Что пошло не так",
            "example": "Сегодня можно лайкнуть не больше 10 событий — возвращайся завтра"
          },
          "message": {
            "type": "string",
            "description": "То же, что `error` (есть у ошибок действий пользователя, код 409)"
          },
          "code": {
            "type": "string",
            "description": "Машинный код ошибки, если есть",
            "example": "VALIDATION_FAILED"
          }
        },
        "description": "Ошибка. Текст в `error` можно показывать пользователю"
      },
...
      "SendSupportMessageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "SupportMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ticketId": {
            "type": "integer",
            "format": "int64"
          },
          "senderType": {
            "type": "string",
            "enum": [
              "USER",
              "OPERATOR"
            ]
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isRead": {
            "type": "boolean"
          }
        }
      },
      "MatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id совпадения",
            "format": "int64",
            "example": 30
          },
          "companionId": {
            "type": "integer",
            "description": "Id напарника",
            "format": "int64",
            "example": 40
          },
          "companionName": {
            "type": "string",
            "description": "Имя напарника",
            "example": "Боря"
          },
          "companionPhoto": {
            "type": "string",
            "description": "Фото напарника"
          },
          "companionAge": {
            "type": "integer",
            "description": "Возраст напарника",
            "format": "int32",
            "example": 21
          },
          "eventId": {
            "type": "integer",
            "description": "Id события",
            "format": "int64",
            "example": 9835
          },
          "eventTitle": {
            "type": "string",
            "description": "Событие",
            "example": "Щелкунчик"
          },
          "eventDate": {
            "type": "string",
            "description": "Дата события",
            "format": "date",
            "example": "2026-10-12"
          },
          "status": {
            "type": "string",
            "description": "NEW — никто не позвал, REQUESTED — приглашение отправлено, ACCEPTED — договорились, чат открыт, DECLINED — отказ",
            "example": "REQUESTED"
          },
          "requestedByMe": {
            "type": "boolean",
            "description": "Приглашение отправил текущий пользователь"
          },
          "chatId": {
            "type": "integer",
            "description": "Id чата — после ACCEPTED",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "Когда нашлось совпадение",
            "format": "date-time"
          }
        },
        "description": "Найденный напарник по событию"
      },
...
      "SendGroupMessageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "GroupChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "senderId": {
            "type": "integer",
            "format": "int64"
          },
          "senderName": {
            "type": "string"
          },
          "senderPhoto": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SendFriendGroupMessageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "FriendGroupChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "senderId": {
            "type": "integer",
            "format": "int64"
          },
          "senderName": {
            "type": "string"
          },
          "senderPhoto": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SendMessageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Текст сообщения",
            "example": "Привет! Идём вместе?"
          }
        },
        "description": "Новое сообщение"
      },
...
      "PageSupportMessageResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportMessageResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "PageGroupChatMessageResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupChatMessageResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "PageFriendGroupChatMessageResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendGroupChatMessageResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "PageChatMessageResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Некорректные данные запроса",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "JOIN доступен с 14 лет",
              "code": "VALIDATION_FAILED"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Нет авторизации (тело пустое) или нет доступа к объекту"
      },
      "NotFound": {
        "description": "Объект не найден или недоступен",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Event not found with id: 42"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Токен из POST /api/auth/login или /api/auth/register",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "maxInitData": {
        "type": "apiKey",
        "description": "window.WebApp.initData из MAX Bridge; подпись проверяется токеном бота",
        "name": "X-Max-Init-Data",
        "in": "header"
      }
    }
  }
}